requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
billiard==3.6.4.0
black==22.1.0
bleach==4.1.0
braceexpand==0.1.7
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
configparser==5.2.0
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
docopt==0.6.2
editdistance==0.6.0
entrypoints==0.3
executing==0.8.2
ffmpeg-python==0.2.0
filelock==3.4.2
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
flask-ipban==1.1.5
Flask-SocketIO==4.3.2
fonttools==4.29.0
frozendict==2.2.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
g2p-en==2.1.0
google-auth==2.5.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
huggingface-hub==0.4.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.10.1
inflect==5.3.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.3.2
kombu==5.2.3
librosa==0.8.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nemo-toolkit[asr]==1.5.1
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.8
numba==0.55.0
numpy==1.19.4
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.10.2
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pooch==1.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==4.3
pyannote.database==4.1.1
pyannote.metrics==3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.1
pytorch-lightning==1.5.9
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.96
shellingham==1.4.0
simplejson==3.17.6
six==1.16.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sox==1.4.1
SQLAlchemy==1.4.31
stack-data==0.1.4
sympy==1.9
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==2.0.0
torch==1.10.1
torch-stft==0.1.4
torchmetrics==0.7.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typer==0.4.0
typing-extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
webrtcvad==2.0.10
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
